+++ YandexMobileAdsExample/app/src/androidTest/java/com/yandex/ads/sample/tests/mediation/StickyBannerLoadTest.kt
@@ -10,11 +10,13 @@ import com.yandex.ads.sample.shared_steps.choiceNetwork
 import com.yandex.ads.sample.shared_steps.goToSection
 import com.yandex.ads.sample.shared_steps.openSampleApp
 import io.github.kakaocup.kakao.screen.Screen.Companion.onScreen
+import org.junit.Ignore
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
 
+@Ignore
 @RunWith(Parameterized::class)
 internal class StickyBannerLoadTest(
     private val networkItem: Class<out StickyBannerScreen.NetworkItem>