+++ changelogs/mediation/bigoads/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to Bigo Ads Adapter for Yandex Mobile Ads Mediation will be documented in this file.
 
+## Version 5.5.1.2
+
+#### Updated
+* Yandex Mobile Ads SDK version 7.18.0
+
 ## Version 5.5.1.1
 
 #### Updated